Handle Route Advertisements
These are sent after the initial handshake events to other nodes and propagate the nodes that are believed to have seen the advertisements
Nodes that receive the advertisement will also repropagate them to their neighbors (excluding the sender).

I also adjust the payload separator for handshake and route advertisements to a semicolon. Since there are a lot of naturally occuring colons in
the route dictionaries, but no obvious semicolon usage.

I can't help but think that these types of messages might be better served as payloads like directives, but that doesn't really fit the model
of directive style messages. Still, I'm not super happy with having to check the payload type using strings in the data receiver