Tweak documentation formatting
Add line breaks into long sample bash commands, so that the code samples
can be read without horizontal scrolling.

Mark the code blocks as such, so that generated HTML is more
semantically correct, and so that language-specific syntax highlighting
may be applied.

Change an unordered list to a definition list, as the latter is more
semantically accurate.

Delete trailing whitespace.

Fix a syntax error. The following should presumably not have whitespace
before the equals sign.

    --recipient =node-b