Use setuptools for dependency mgmt & for building
This project uses poetry for dependency management and as a build
backend when doing development, and setuptools for those tasks when
targeting production (e.g. building RPMs). There are several downsides
to this situation:

*   Tooling duplication creates extra work for everyone who touches this
    project, as those people must learn when to use which tool, and
    possibly master both.
*   Developers will typically use poetry, but other users (including
    packagers and end users) will typically use setuptools. This
    disparity creates distance between developers and everyone else.
*   Differences between the poetry and setuptools toolchains are highly
    likely to arise. Attempting to support both equally will likely
    fail, and any discrepancies are bugs.

This commit drops poetry, and uses exclusively setuptools.

Poetry has an excellent dependency resolver and makes managing abstract
and concrete dependencies a breeze. In contrast, pip doesn't have a
proper dependency resolver, and managing concrete dependencies with a
requirements file is doable but more of a chore.  (AFAICT, setuptools
doesn't have a dependency resolver.)

Furthermore, this project made use of `poetry.lock` to manage concrete
dependencies, but (at this time!) has no corresponding pip requirements
or constraints files.

Consequently, dropping poetry may be painful for developers. However,
the author of this commit believes that the correct way to deal with
that pain is for everyone to settle on and make the best use possible of
a single set of tools, not by using two duplicate tools.

This commit is conservative. Some changes were intentionally minimal, at
the expense of possibly leaving some technical debt in place:

*   The `poetry.lock` file was not replaces with a setuptools
    requirements or constraints file.
*   The makefile's `shell` target was left in place.
*   The makefile's `DIST_PYTHON` variable was left in place.
*   Pinned dependencies were left in `setup.py`, i.e. some requirements
    in `setup.py` make use of `==`.

See: https://github.com/project-receptor/receptor/issues/94