Fix several docstring warnings
Building the docs produces the following warnings:

    …/receptor/controller.py:docstring of receptor.controller.Controller.recv:1: WARNING: Inline emphasis start-string without end-string.
    …/receptor/messages/framed.py:docstring of receptor.messages.framed:11: WARNING: Definition list ends without a blank line; unexpected unindent.
    …/receptor/messages/framed.py:docstring of receptor.messages.framed:16: WARNING: Definition list ends without a blank line; unexpected unindent.
    …/docs/source/modules.rst: WARNING: document isn't included in any toctree

Fix the first three. The solution to the fourth isn't immediately
obvious to me, and fixing it requires a knowledge of the developers'
intent.

Issues like this can be more easily detected by building docs as part of
CI and setting `nitpicky = True` in `docs/conf.py`.